+++ accounting/src/__tests__/mockserver/internal/handler/pathpoststoreorder.go
@@ -25,7 +25,7 @@ func pathPostStoreOrder(dir *logging.HTTPFileDirectory, rt *tracking.RequestTrac
 		case "placeOrder[0]":
 			dir.HandlerFunc("placeOrder", testPlaceOrderPlaceOrder0)(w, req)
 		default:
-			http.Error(w, "Unknown test: "+test, http.StatusBadRequest)
+			http.Error(w, fmt.Sprintf("Unknown test: %s[%d]", test, count), http.StatusBadRequest)
 		}
 	}
 }
@@ -51,7 +51,7 @@ func testPlaceOrderPlaceOrder0(w http.ResponseWriter, req *http.Request) {
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
-	respBody := &components.Order{
+	var respBody *components.Order = &components.Order{
 		ID:       types.Int64(10),
 		PetID:    types.Int64(198772),
 		Quantity: types.Int(7),