+++ accounting/src/__tests__/mockserver/internal/handler/pathputpet.go
@@ -25,7 +25,7 @@ func pathPutPet(dir *logging.HTTPFileDirectory, rt *tracking.RequestTracker) htt
 		case "petsStoreMonday[0]":
 			dir.HandlerFunc("petsStoreMonday", testPetsStoreMondayPetsStoreMonday0)(w, req)
 		default:
-			http.Error(w, "Unknown test: "+test, http.StatusBadRequest)
+			http.Error(w, fmt.Sprintf("Unknown test: %s[%d]", test, count), http.StatusBadRequest)
 		}
 	}
 }
@@ -51,7 +51,7 @@ func testPetsStoreMondayPetsStoreMonday0(w http.ResponseWriter, req *http.Reques
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
-	respBody := &components.Pet{
+	var respBody *components.Pet = &components.Pet{
 		ID:   types.Int64(10),
 		Name: "doggie",
 		Category: &components.Category{