+++ accounting/src/funcs/storeGetInventory.ts
@@ -9,6 +9,7 @@ import { compactMap } from "../lib/primitives.js";
 import { RequestOptions } from "../lib/sdks.js";
 import { extractSecurity, resolveGlobalSecurity } from "../lib/security.js";
 import { pathToFunc } from "../lib/url.js";
+import { AccountingSDKError } from "../models/errors/accountingsdkerror.js";
 import {
   ConnectionError,
   InvalidRequestError,
@@ -17,8 +18,9 @@ import {
   UnexpectedClientError,
 } from "../models/errors/httpclienterrors.js";
 import * as errors from "../models/errors/index.js";
-import { SDKError } from "../models/errors/sdkerror.js";
+import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
+import { APICall, APIPromise } from "../types/async.js";
 import { Result } from "../types/fp.js";
 
 /**
@@ -27,22 +29,50 @@ import { Result } from "../types/fp.js";
  * @remarks
  * Returns a map of status codes to quantities
  */
-export async function storeGetInventory(
+export function storeGetInventory(
   client: AccountingSDKCore,
   options?: RequestOptions,
-): Promise<
+): APIPromise<
   Result<
     { [k: string]: number },
     | errors.ApiErrorUnauthorized
     | errors.ApiErrorNotFound
-    | SDKError
-    | SDKValidationError
-    | UnexpectedClientError
-    | InvalidRequestError
+    | AccountingSDKError
+    | ResponseValidationError
+    | ConnectionError
     | RequestAbortedError
     | RequestTimeoutError
-    | ConnectionError
+    | InvalidRequestError
+    | UnexpectedClientError
+    | SDKValidationError
   >
+> {
+  return new APIPromise($do(
+    client,
+    options,
+  ));
+}
+
+async function $do(
+  client: AccountingSDKCore,
+  options?: RequestOptions,
+): Promise<
+  [
+    Result<
+      { [k: string]: number },
+      | errors.ApiErrorUnauthorized
+      | errors.ApiErrorNotFound
+      | AccountingSDKError
+      | ResponseValidationError
+      | ConnectionError
+      | RequestAbortedError
+      | RequestTimeoutError
+      | InvalidRequestError
+      | UnexpectedClientError
+      | SDKValidationError
+    >,
+    APICall,
+  ]
 > {
   const path = pathToFunc("/store/inventory")();
 
@@ -55,7 +85,8 @@ export async function storeGetInventory(
   const requestSecurity = resolveGlobalSecurity(securityInput);
 
   const context = {
-    baseURL: options?.serverURL ?? "",
+    options: client._options,
+    baseURL: options?.serverURL ?? client._baseURL ?? "",
     operationID: "getInventory",
     oAuth2Scopes: [],
 
@@ -74,10 +105,11 @@ export async function storeGetInventory(
     baseURL: options?.serverURL,
     path: path,
     headers: headers,
+    userAgent: client._options.userAgent,
     timeoutMs: options?.timeoutMs || client._options.timeoutMs || -1,
   }, options);
   if (!requestRes.ok) {
-    return requestRes;
+    return [requestRes, { status: "invalid" }];
   }
   const req = requestRes.value;
 
@@ -88,7 +120,7 @@ export async function storeGetInventory(
     retryCodes: context.retryCodes,
   });
   if (!doResult.ok) {
-    return doResult;
+    return [doResult, { status: "request-error", request: req }];
   }
   const response = doResult.value;
 
@@ -100,23 +132,24 @@ export async function storeGetInventory(
     { [k: string]: number },
     | errors.ApiErrorUnauthorized
     | errors.ApiErrorNotFound
-    | SDKError
-    | SDKValidationError
-    | UnexpectedClientError
-    | InvalidRequestError
+    | AccountingSDKError
+    | ResponseValidationError
+    | ConnectionError
     | RequestAbortedError
     | RequestTimeoutError
-    | ConnectionError
+    | InvalidRequestError
+    | UnexpectedClientError
+    | SDKValidationError
   >(
     M.json(200, z.record(z.number().int())),
     M.jsonErr(401, errors.ApiErrorUnauthorized$inboundSchema),
     M.jsonErr(404, errors.ApiErrorNotFound$inboundSchema),
     M.fail("4XX"),
     M.fail("5XX"),
-  )(response, { extraFields: responseFields });
+  )(response, req, { extraFields: responseFields });
   if (!result.ok) {
-    return result;
+    return [result, { status: "complete", request: req, response }];
   }
 
-  return result;
+  return [result, { status: "complete", request: req, response }];
 }